+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 apidoc extension Change Log
 2.0.6 under development
 -----------------------
 
+- Bug #5: Enable display of deprecated information for methods, properties, constants and events (cebe)
 - Bug #62: Make `@inheritdoc` tag more robust (cebe, sasha-ch)
 - Bug #68: Fixed crash on empty type in PHPdoc (cebe, itnelo)
 - Bug #76: Fixed broken links with external urls (CedricYii)